requirements.txt
selenium
python-dotenv
ollama
